rpi-eeprom: new recipe
This recipe is insipred by the one from Balena:
https://github.com/balena-os/balena-raspberrypi/blob/master/layers/\
  meta-balena-raspberrypi/recipes-bsp/rpi-eeprom/rpi-eeprom_git.bb

The one from Balena, however, is suited to their custom update method.
The goal of this rpi-eeprom recipe is to ship the eeprom update scripts
and firmware files in a similar fashion as it is done in the Raspbian.

The rpi-eeprom package contains firmware and update scripts.
The rpi-eeprom-config package contains optional config script, which
depends on Python so it is packaged separately.

To update EEPROM firmware simply run "rpi-eeprom-update" -a and reboot.

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>