userland: add userlandtools recipe
The goal is that the userlandtools recipe contains some tools coming
from the userland, which are required to perform the EEPROM update. At
the same time, we do not want to use the full userland, as we prefer to
use the mainline mesa userspace GPU libraries.

As part of this work, userland.inc file was introduced. It contains
common parts for the userland and userlandtools recipes.

The userlandtools recipe is inspired by the one from Balena:
https://github.com/balena-os/balena-raspberrypi/blob/master/layers/\
  meta-balena-raspberrypi/recipes-extended/userlandtools/userlandtools.bb

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>